main/requirements.txt
# 个人环境配置，部分包在此项目中未使用

albucore==0.0.17
albumentations==1.4.11
annotated-types==0.7.0
appdirs==1.4.4
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
cython==3.1.2
dill==0.3.8
efficientnet-pytorch==0.7.1
einops==0.8.1
eval-type-backport==0.2.2
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
grad-cam==1.4.8
huggingface-hub==0.33.0
idna==3.10
imageio==2.35.1
importlib-resources==6.4.5
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.5
lazy-loader==0.4
matplotlib==3.7.2
networkx==3.1
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prettytable==3.11.0
psutil==7.0.0
py-cpuinfo==9.0.0
pycocotools==2.0
pydantic==2.10.6
pydantic-core==2.27.2
pyparsing==3.0.9
python-dateutil==2.9.0.post0
pytorch-wavelets==1.3.0
pytz==2025.2
pywavelets==1.4.1
PyYAML==6.0.2
requests==2.32.4
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.3
thop==0.1.1-2209072238
threadpoolctl==3.6.0
tidecv==1.0.1
tifffile==2023.7.10
timm==1.0.7
torch==2.8.0.dev20250617+cu128
torchaudio==2.8.0.dev20250618+cu128
torchvision==0.23.0.dev20250618+cu128
tqdm==4.67.1
ttach==0.0.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.20.2
